% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_dossier}
\alias{render_dossier}
\title{Process and render a dossier}
\usage{
render_dossier(
  input = "dossier.md",
  output = c("dossier.docx", "dossier.pdf"),
  cleanup = TRUE
)
}
\arguments{
\item{input}{Input markdown file.}

\item{output}{A vector of desired outputs.}

\item{cleanup}{Delete the processed markdown file?}
}
\value{
One or more rendered dossier files. Defaults to docx and pdf.
Can also do html or a single format.
}
\description{
Process and render a dossier
}
\examples{
\dontrun{
file.copy(system.file("dossier.md", package = "resdown"), ".", overwrite = FALSE)
render_dossier()
}
}
