% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resdown-package.R
\docType{package}
\name{resdown-package}
\alias{resdown}
\alias{resdown-package}
\title{resdown: Render DFO Research Scientist Dossiers in Markdown with R}
\description{
Unofficial Markdown template and associated R functions to produce dossiers for Fisheries and Oceans Canada Research Scientists under the Career Progression Management Framework.
}
\author{
\strong{Maintainer}: Sean C. Anderson \email{sean@seananderson.ca} (\href{https://orcid.org/0000-0001-9563-1937}{ORCID})

}
\keyword{internal}
